+++ app/build.gradle.kts
@@ -42,6 +42,10 @@ android {
         targetCompatibility = JavaVersion.VERSION_1_8
     }
 
+    buildFeatures {
+        viewBinding = true
+    }
+
 }
 
 dependencies {
