+++ app/src/main/java/com/example/todoapp/presentation/ui/screen/main_screen/adapter/TodoAdapter.kt
@@ -1,22 +1,17 @@
-package com.example.todoapp.presentation.ui.main_screen.adapter
+package com.example.todoapp.presentation.ui.screen.main_screen.adapter
 
 import android.graphics.Color
 import android.graphics.Paint
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.CheckBox
-import android.widget.LinearLayout
-import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.example.todoapp.R
 import com.example.todoapp.data.model.Importance
 import com.example.todoapp.data.model.TodoItem
-import com.example.todoapp.data.repository.TodoItemsRepository
 import java.time.LocalDateTime
-
+import com.example.todoapp.databinding.TodoItemBinding
 
 class TodoAdapter(
     private val onTasksChangedListener: OnTaskChangeListener,
@@ -35,9 +30,8 @@ class TodoAdapter(
 
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TodoViewHolder {
-        val itemView =
-            LayoutInflater.from(parent.context).inflate(R.layout.todo_item, parent, false)
-        return TodoViewHolder(itemView)
+        val binding = TodoItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        return TodoViewHolder(binding)
     }
 
     override fun onBindViewHolder(holder: TodoViewHolder, position: Int) {
@@ -46,24 +40,24 @@ class TodoAdapter(
     }
 
 
-    class TodoViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-        private val tvTodoText: TextView = itemView.findViewById(R.id.mainText)
-        private val checkBox: CheckBox = itemView.findViewById(R.id.checkboxCompleted)
-        private val tvImportance: TextView = itemView.findViewById(R.id.importanceFlag)
-        private val editClickArea: LinearLayout = itemView.findViewById(R.id.editTaskClickArea)
+    class TodoViewHolder(binding: TodoItemBinding) : RecyclerView.ViewHolder(binding.root) {
+        private val tv = binding.mainText
+        private val checkBox = binding.checkboxCompleted
+        private val tvImportance = binding.importanceFlag
+        private val editClickArea = binding.editTaskClickArea
 
         fun bind(
             todoItem: TodoItem,
             onTasksChangedListener: OnTaskChangeListener,
             onTaskPressListener: OnTaskPressListener,
         ) {
-            tvTodoText.text = todoItem.text
+            tv.text = todoItem.text
             checkBox.setOnCheckedChangeListener(null)
             checkBox.isChecked = todoItem.done
-            updateTextDecoration(tvTodoText, todoItem.done, todoItem.deadline)
+            updateTextDecoration(todoItem.done, todoItem.deadline)
 
             checkBox.setOnCheckedChangeListener { _, isChecked ->
-                updateTextDecoration(tvTodoText, isChecked, todoItem.deadline)
+                updateTextDecoration(isChecked, todoItem.deadline)
                 onTasksChangedListener.onTaskChanged(todoItem.id)
             }
 
@@ -71,36 +65,40 @@ class TodoAdapter(
                 onTaskPressListener.onTaskPressed(todoItem.id)
             }
 
-            when (todoItem.importance) {
+            setUpImportance(todoItem.importance)
+        }
+
+        private fun setUpImportance(importance: Importance) {
+            val context = itemView.context
+            when (importance) {
                 Importance.LOW -> {
-                    tvImportance.text = ""
+                    tvImportance.text = context.getString(R.string.empty_string)
                     tvImportance.setTextColor(Color.BLACK)
                 }
 
                 Importance.MEDIUM -> {
-                    tvImportance.text = "!"
+                    tvImportance.text = context.getString(R.string.medium_importance_symbol)
                     tvImportance.setTextColor(Color.BLACK)
                 }
 
                 Importance.HIGH -> {
-                    tvImportance.text = "!!"
+                    tvImportance.text = context.getString(R.string.high_importance_symbol)
                     tvImportance.setTextColor(Color.RED)
                 }
             }
         }
 
 
-        private fun updateTextDecoration(tv: TextView, done: Boolean, deadline: LocalDateTime?) {
+        private fun updateTextDecoration(done: Boolean, deadline: LocalDateTime?) {
             if (done) {
                 tv.paintFlags = tv.paintFlags or Paint.STRIKE_THRU_TEXT_FLAG
+                tv.setTextColor(Color.GRAY)
+            } else if (deadline != null && deadline <= LocalDateTime.now()) {
+                tv.paintFlags = tv.paintFlags and Paint.STRIKE_THRU_TEXT_FLAG.inv()
+                tv.setTextColor(Color.RED)
             } else {
                 tv.paintFlags = tv.paintFlags and Paint.STRIKE_THRU_TEXT_FLAG.inv()
-            }
-
-            if (!done && deadline != null && deadline <= LocalDateTime.now()) {
-                tvTodoText.setTextColor(Color.RED)
-            } else {
-                tvTodoText.setTextColor(Color.BLACK)
+                tv.setTextColor(Color.BLACK)
             }
         }
     }
