+++ app/src/main/res/layout/activity_main.xml
@@ -1,26 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    android:layout_height="match_parent">
 
     <FrameLayout
         android:id="@+id/fragment_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginRight="20dp"
-        android:layout_marginBottom="40dp"
-        android:backgroundTint="@color/blue"
-        android:src="@drawable/baseline_add_24"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentBottom="true"
-        app:fabSize="normal"
-        app:tint="@color/white"
-        app:shapeAppearanceOverlay="@style/fab_circle"
-        android:contentDescription="add_task_button" />
 </RelativeLayout>
