+++ app/src/main/res/layout/fragment_main.xml
@@ -93,7 +93,20 @@
         app:layout_constraintVertical_bias="0.0" />
 
 
-
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:id="@+id/fab"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="16dp"
+        android:layout_marginBottom="40dp"
+        android:backgroundTint="@color/blue"
+        android:contentDescription="@string/add_task_button_description"
+        android:src="@drawable/baseline_add_24"
+        app:fabSize="normal"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:shapeAppearanceOverlay="@style/fab_circle"
+        app:tint="@color/white" />
